Add galleries based on Filer folders
This is more likely to be useful for my applications - rather than
assigning individual images to a gallery, this allows assigning an
entire folder as uploaded with django-filer.

To support this, thumbnails had to be modified - rather than the size
being specified by a function in GalleryFolder, it is now done with
a template tag. This will have to be watched for performance
implications but I'm not concerned about that at the moment.